made return value of compute_objective_function double as opposed to float. this should avoid problems with accumulating lots of numbers of the same magnitude. This meant changing support functions as well.

+++ src/include/local/stir/recon_buildblock/PatlakObjectiveFunctionFromDynamicProjectionData.h
@@ -85,7 +85,7 @@ public  RegisteredParsingObject<PatlakObjectiveFunctionFromDynamicProjectionData
 							  const TargetT &current_estimate, 
 							  const int subset_num); 
 
-  virtual float 
+  virtual double
     actual_compute_objective_function_without_penalty(const TargetT& current_estimate,
 						      const int subset_num);
